42-OpenStep←X2-Jaguar

+++ Soulver/SVRSolverSolutionTagger.m
@@ -387,7 +387,7 @@ NSSet *SVRSolverSolutionTaggerSetAddSub   = nil;
     case SVRSolverOperatorLog:
       return [rhs SVR_decimalNumberByLogarithmWithBase:lhs withBehavior:ohBehave];
     default:
-      XPLogRaise1(@"__solveWithOperatorUnknown:%d", operator);
+      XPLogRaise1(@"__solveWithOperatorUnknown:%d", (int)operator);
       return nil;
   }
 }
